adds integer min max type instead of global min max type to be more specific and fastly reusable - adds default most common used attributes
