+++ admin-dev/themes/new-theme/scss/components/_grid.scss
@@ -91,7 +91,7 @@ table .column-filters td:last-child .grid-reset-button {
           .min-field,
           .max-field {
             display: inline-block;
-            width: 50px;
+            width: 60px;
           }
         }
       }
