+++ src/PrestaShopBundle/Resources/views/Admin/TwigTemplateForm/bootstrap_4_layout.html.twig
@@ -781,7 +781,7 @@
   </small>
 {% endblock text_with_recommended_length_widget %}
 
-{% block min_max_filter_widget %}
+{% block integer_min_max_filter_widget %}
   {{ form_widget(form['min_field'], { attr: {class: 'min-field'}}) }}
   {{ form_widget(form['max_field'], { attr: {class: 'max-field'}}) }}
 {% endblock %}
